39: Mesh renders two different meshes simultaneously Bug was caused by instance draw call manually setting 100. Corrected to only draw n instances where n is the size of the commands.

+++ Source/ScarletEngineProject/ScarletEngine/Rendering/IndexBuffer.cpp
@@ -6,14 +6,14 @@
 namespace Scarlet
 {
 
-IndexBuffer::IndexBuffer(const uint32 count)
+IndexBuffer::IndexBuffer(const size_t count)
 {
     glCreateBuffers(1, &mId);
     Bind();
     glBufferData(GL_ELEMENT_ARRAY_BUFFER, static_cast<long long>(count * sizeof(uint32)), nullptr, GL_DYNAMIC_DRAW);
 }
 
-IndexBuffer::IndexBuffer(const uint32* indices, const uint32 count)
+IndexBuffer::IndexBuffer(const uint32* indices, const size_t count)
 {
     glCreateBuffers(1, &mId);
     Bind();
